# optimize/fix DefaultValueConsistencyAnalyzer