With ImmutabilityAnalyzer as reference (given run variability):

| Analyzer | Before | After |
| -- | -- | -- |
| ImmutabilityAnalyzer | 11.804s | 11.681s |
| DefaultValueConsistencyAnalyzer | 4.261s | 1.340s |