@omsmith 2023-11-20
Ref: https://github.com/Brightspace/lms/pull/43590